pyproject.toml
[project]
description = "Default template for PDM package"
name = "hycBot"
readme = "README.md"
requires-python = "==3.12.*"
version = "0.1.0"
authors = [
    {name = "2c2048d2", email = "dyq233@qq.com"},
    {name = "lordmystery", email = "1517076134@qq.com"},
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "certifi==2024.2.2",
    "click==8.1.7",
    "dnspython==2.6.1",
    "email-validator==2.1.1",
    "fastapi==0.111.0",
    "fastapi-cli==0.0.4",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "idna==3.7",
    "Jinja2==3.1.4",
    "loguru==0.7.2",
    "markdown-it-py==3.0.0",
    "MarkupSafe==2.1.5",
    "mdurl==0.1.2",
    "msgpack==1.0.8",
    "multidict==6.0.5",
    "nonebot-adapter-onebot==2.4.3",
    "nonebot2>=2.3.2",
    "orjson==3.10.3",
    "pydantic>=2.7.2",
    "pydantic-core==2.18.3",
    "Pygments==2.18.0",
    "pygtrie==2.5.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "PyYAML==6.0.1",
    "rich==13.7.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "starlette==0.37.2",
    "typer==0.12.3",
    "typing-extensions==4.12.0",
    "ujson==5.10.0",
    "uvicorn==0.30.0",
    "uvloop==0.19.0",
    "watchfiles==0.22.0",
    "websockets==12.0",
    "yarl==1.9.4",
    "requests>=2.32.3",
    "lxml>=5.2.2",
    "nonebot-plugin-htmlrender>=0.3.2",
    "pymysql>=1.1.1",
    "arrow>=1.3.0",
    "nonebot-adapter-console>=0.6.0",
    "mysql-connector-python>=9.0.0",
    "sqlalchemy>=2.0.31",
]
license = {text = "MIT"}

[tool.nonebot]
adapters = [
  {name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11"},
]
builtin_plugins = ["echo"]
plugin_dirs = ["plugins"]
plugins = ["nonebot_plugin_htmlrender"]

[tool.pdm]
distribution = false
